Improved perf to add measure method with lambda.
