+++ slider/platformio.ini
@@ -24,7 +24,7 @@ build_flags =
 	-D IS_SIMULATOR=TRUE
 	-D TEST_NETWORK=TRUE
 	-D IS_RECEIVER=TRUE
-	-D DEBUG_PERF=TRUE
+	-D LOG_PERFORMANCE=TRUE
 	-I ../slider
 debug_build_flags = -O0 -g -ggdb
 build_type = debug
