--- slider/src/core/perf/cpuUsage.h
@@ -1,25 +1,0 @@
-#ifndef CPUUSAGE_H
-#define CPUUSAGE_H
-
-#include "perf.h"
-
-namespace Performance
-{
-/*
-    struct CpuUsage
-    {
-    public:
-        struct Tag { static constexpr const char* Name { "Cpu usage" }; };
-        static constexpr const char* Unit { "%" };
-        void Start();
-        uint8_t GetValue() const;
-    private:
-
-    };
-
-    template class Sampler<CpuUsage::Tag, CpuUsage, uint8_t>;
-    using CpuUsageSampler = Sampler<CpuUsage::Tag, CpuUsage, uint8_t>;
-*/
-}
-
-#endif
