+++ slider/src/core/perf/perf.cpp
@@ -1,6 +1,30 @@
-#include "cpuUsage.h"
+#include "perf.h"
+
+#ifdef LOG_PERFORMANCE
+
+#ifdef ARDUINO_ARCH_ESP32 
+#include "esp_timer.h"
+#endif
 
 using namespace Performance;
+
+void CpuTime::Start()
+{
+    #ifdef ARDUINO_ARCH_ESP32
+    m_StartMicroseconds = esp_timer_get_time();
+    #else
+    m_StartMicroseconds = micros();
+    #endif
+}
+
+uint64_t CpuTime::GetValue() const
+{
+    #ifdef ARDUINO_ARCH_ESP32
+    return (esp_timer_get_time() - m_StartMicroseconds);
+    #else
+    return (micros() - m_StartMicroseconds);
+    #endif
+}
 
 /*
 #include "freertos/FreeRTOS.h"
@@ -52,3 +76,5 @@ void CpuUsage::Start()
     }
 }
 */
+
+#endif //LOG_PERFORMANCE
