+++ slider/src/debug.h
@@ -39,20 +39,6 @@
 #define LogDebug(...) ;
 #endif
 
-#ifdef DEBUG_PERF
-#define LogPerf(...) Debug::Logger().Log(__VA_ARGS__)
-#define SAMPLER_START(sampler) sampler.Start();
-#define SAMPLER_END(sampler) sampler.End();
-#define CREATE_SAMPLER(type, name, frequency) static Performance::type name(frequency);
-#define MEASURE(sampler, func) sampler.Measure([]()func);
-#else
-#define LogPerf(...) ;
-#define SAMPLER_START(sampler) ;
-#define SAMPLER_END(sampler) ;
-#define CREATE_SAMPLER(type, name, frequency) ;
-#define MEASURE(sampler, func) func;
-#endif
-
 namespace Debug
 {
     struct Logger
