+++ slider/src/main.cpp
@@ -18,6 +18,7 @@
 #include "src/core/app/appBase.h"
 #include "src/brain/appConfig.h"
 #include "debug.h"
+#include "src/core/perf/perf.h"
 #include <memory>
 
 using namespace Slider;
@@ -42,10 +43,7 @@ static AppConfig CreateConfig()
 
 static auto app = Core::AppCreator<AppConfig>::Create(CreateConfig());
 
-#ifdef DEBUG_PERF
-#include "src/core/perf/cpuTime.h"
-static Performance::CpuTimeSampler cpuSampler(200u);
-#endif
+CREATE_SAMPLER(CpuTimeSampler, cpuSampler, 200u);
 
 void setup()
 {
@@ -58,9 +56,10 @@ void setup()
 
 void loop()
 {
-    SAMPLER_START(cpuSampler);
-    app->Update();
-    SAMPLER_END(cpuSampler);
+    MEASURE(cpuSampler, 
+    {
+        app->Update();
+    });
 
     #ifdef IS_SIMULATOR //somehow this makes the timing more accurate...
     delay(10);
